requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic-settings
groq==0.4.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow
pytesseract==0.3.10
aiofiles==23.2.0
httpx==0.25.2